DSTEW-1327 New controller for testing versioning
