+++ claims-data/api/open-api-specification.yml
@@ -336,7 +336,6 @@ paths:
       x-spring-paginated: true
       tags:
         - Claims
-        - v2
       summary: Get claims
       description: Get claims for an office, filtering on certain fields.
       parameters:
@@ -441,7 +440,6 @@ paths:
       x-spring-paginated: true
       tags:
         - Claims
-        - v1
       summary: Get claims
       description: Get claims for an office, filtering on certain fields.
       parameters:
@@ -780,7 +778,6 @@ paths:
       operationId: getClaim
       tags:
         - Claims
-        - v1
       summary: Get a specific claim for a submission
       description: Returns the details of a specific claim associated with a submission.
       parameters:
@@ -861,7 +858,6 @@ paths:
       operationId: getClaimV2
       tags:
         - Claims
-        - v2
       summary: Get a specific claim for a submission
       description: Returns the details of a specific claim associated with a submission.
       parameters:
