+++ claims-data/service/src/main/resources/application.yml
@@ -22,6 +22,7 @@ spring:
   jackson:
     deserialization:
       fail-on-unknown-properties: true
+
   jpa:
     show-sql: false
     generate-ddl: false
